Remove static method functions. Expand to most applicable method directly instead of a MATCH form.
Static method functions are removed as CCL (and probably other
implementations) "forget" the GENSYM'd symbols and the corresponding
function definitions. Static method functions will have to be named
with interned symbols.

Instead of expanding to a match form the types of the arguments are
determined directly and the most applicable method is chosen directly
instead of expanding to a match form. This is done as most compilers
are not "smart enough" to remove the unused branches, rendering the
whole optimization useless.